Separate PartingGift from Adon Fox
