+++ Assets/Scripts/Model/Content/SecondEdition/Pilots/Rebel/ASF01BWing/AdonFoxBattleOverEndor.cs
@@ -97,59 +97,4 @@ namespace Abilities.SecondEdition
             dieCount++;
         }
     }
-
-    public class PartingGiftAbility : GenericAbility
-    {
-        public override void ActivateAbility()
-        {
-            HostShip.OnShipIsDestroyed += RegisterAbility;
-        }
-
-        public override void DeactivateAbility()
-        {
-            HostShip.OnShipIsDestroyed -= RegisterAbility;
-        }
-
-        private void RegisterAbility(GenericShip ship, bool isFled)
-        {
-            if (!isFled)
-            {
-                RegisterAbilityTrigger(TriggerTypes.OnShipIsDestroyed, UseAbility);
-            }        
-        }
-
-        private void UseAbility(object sender, EventArgs e)
-        {
-            HostShip.OnGetAvailableBombDropTemplatesTwoConditions += GetBombTemplates;
-            BombsManager.RegisterBombDropTriggerIfAvailable(HostShip, TriggerTypes.OnAbilityDirect);
-            Triggers.ResolveTriggers(TriggerTypes.OnAbilityDirect, Triggers.FinishTrigger);
-        }
-
-        private void GetBombTemplates(List<ManeuverTemplate> availableTemplates, GenericUpgrade upgrade)
-        {
-            HostShip.OnGetAvailableBombDropTemplatesTwoConditions -= GetBombTemplates;
-            availableTemplates.Add(new ManeuverTemplate(ManeuverBearing.Bank, ManeuverDirection.Left, ManeuverSpeed.Speed1, isBombTemplate: true));
-            availableTemplates.Add(new ManeuverTemplate(ManeuverBearing.Bank, ManeuverDirection.Right, ManeuverSpeed.Speed1, isBombTemplate: true));
-        }
-    }
-}
-
-namespace UpgradesList.SecondEdition
-{
-    public class PartingGift : GenericUpgrade
-    {
-        public PartingGift() : base()
-        {
-            UpgradeInfo = new UpgradeCardInfo(
-                "Parting Gift",
-                UpgradeType.Talent,
-                cost: 0,
-                abilityType: typeof(PartingGiftAbility)
-            );
-
-            IsHidden = true;
-
-            ImageUrl = HostShip != null? HostShip.ImageUrl : "https://infinitearenas.com/xw2/images/quickbuilds/adonfox-battleoverendor.png";
-        }
-    }
 }
