+++ Assets/Scripts/Model/Content/SecondEdition/Upgrades/Talent/PartingGift.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6a62a42c89338ad49babfe2d1382ca45
